Add menu, split GUI module into 3
