+++ src/parsagon/gui/__init__.py
@@ -1,0 +1,2 @@
+from .controller import GUIController
+from .window import GUIWindow
