+++ src/parsagon/gui/window.py
@@ -1,9 +1,9 @@
-import contextlib
 import os
+from datetime import datetime
 from threading import Condition
 
-from PyQt6.QtCore import QThread, pyqtSignal, pyqtSlot, QEventLoop, QSize, QTimer
 from PyQt6.QtCore import Qt
+from PyQt6.QtCore import pyqtSlot, QSize, QTimer
 from PyQt6.QtGui import (
     QMovie,
     QKeyEvent,
@@ -29,19 +29,15 @@ from PyQt6.QtWidgets import (
     QMessageBox,
 )
 
-from parsagon.exceptions import APIException
-from parsagon.settings import get_graphic, get_save_api_key_interactive
+from parsagon.gui.controller import GUIController
+from parsagon.gui.menu import MenuManager
+from parsagon.settings import get_graphic
 
 message_padding_constant = 0
 message_width_proportion = 0.85
 is_windows = os.name == "nt"
 font_size_messages = 13 if not is_windows else 10
 font_size_input = 15 if not is_windows else 11
-
-
-class ResultContainer:
-    def __init__(self):
-        self.value = None
 
 
 try:
@@ -51,99 +47,6 @@ try:
     windll.shell32.SetCurrentProcessExplicitAppUserModelID(app_id)
 except ImportError:
     pass
-
-
-class GUIController(QThread):
-    """
-    The GUI controller is a background thread that runs the CLI if the GUI is enabled.  It then delegates all interaction between the CLI code and the GUI.  It is designed as a singleton so that references to the controller don't have to be passed around.
-    """
-
-    _instance = None
-
-    update_text_signal = pyqtSignal(str, str, str)  # Include text color
-    request_input_signal = pyqtSignal()
-    set_loading_signal = pyqtSignal(bool)
-    show_progress_signal = pyqtSignal(bool)
-    set_progress_title_signal = pyqtSignal(str)
-    set_progress_signal = pyqtSignal(int)
-    execute_on_main_thread_signal = pyqtSignal(object, QEventLoop, ResultContainer)
-
-    @classmethod
-    def shared(cls):
-        if cls._instance is None:
-            raise RuntimeError("GUIController has not been initialized")
-        return cls._instance
-
-    def __init__(self, condition, parent=None):
-        super().__init__(parent)
-        self.timer = None
-        self.condition = condition
-        self.current_input = None
-        self.progress_total = None
-        assert self.__class__._instance is None, "GUIController is a singleton"
-        self.__class__._instance = self
-
-    def run(self):
-        self.__class__._instance._instance = self
-        while True:
-            retry = False
-            try:
-                from parsagon.assistant import assist
-                from parsagon.settings import get_graphic, get_api_key
-
-                _ = get_api_key(interactive=True)
-                assist(True)
-            except Exception as e:
-                if isinstance(e, (KeyboardInterrupt, SystemExit)):
-                    raise
-                else:
-                    from parsagon.print import error_print
-
-                    error_print(str(e))
-                    error_print(
-                        "Parsagon encountered an error.  Please restart the application to continue.  You may want to copy any messages above you want to save."
-                    )
-
-                    if isinstance(e, APIException) and e.status_code == 401:
-                        get_save_api_key_interactive()
-                        retry = True
-            if not retry:
-                break
-
-    def input(self, prompt):
-        if prompt:
-            self.print(prompt)
-        self.request_input_signal.emit()
-        with self.condition:
-            self.condition.wait()
-            return self.current_input
-
-    def print(self, text, color="black", background="#CBCED2"):
-        html = text.replace("\n", "<br>")
-        self.update_text_signal.emit(html, color, background)
-
-    @contextlib.contextmanager
-    def spinner(self):
-        self.set_loading_signal.emit(True)  # Show spinner
-        try:
-            yield
-        finally:
-            self.set_loading_signal.emit(False)  # Hide spinner
-
-    def show_progress(self, show_progress, text=""):
-        self.show_progress_signal.emit(show_progress)
-        if show_progress:
-            self.set_progress_title_signal.emit(text)
-
-    def set_progress(self, progress_iteration_num):
-        self.set_progress_signal.emit(int(progress_iteration_num / self.progress_total * 100))
-
-    def on_main_thread(self, callback):
-        loop = QEventLoop()
-        result_container = ResultContainer()
-        self.execute_on_main_thread_signal.emit(callback, loop, result_container)
-        loop.exec()  # Block until loop.quit() is called in the main thread
-        return result_container.value  # Return the result stored in the container
 
 
 class CustomTextEdit(QTextEdit):
@@ -260,16 +163,8 @@ class GUIWindow(QMainWindow):
         self.central_widget.setLayout(main_layout)
         self.setCentralWidget(self.central_widget)
 
-        # Create the menu bar
-        menu_bar = self.menuBar()
-
-        # Create a menu for the app name
-        app_menu = menu_bar.addMenu("Parsagon")
-
-        # Create an "About" action
-        about_action = QAction("About", self)
-        app_menu.addAction(about_action)
-        about_action.triggered.connect(self.show_about_dialog)
+        self.menu_manager = MenuManager(self)
+        self.menu_manager.make_menu()
 
         # Signals
         self.controller = GUIController(self.condition)
@@ -279,9 +174,9 @@ class GUIWindow(QMainWindow):
         self.controller.set_progress_signal.connect(self.set_progress)
         self.controller.update_text_signal.connect(self.update_text_ui)
         self.controller.request_input_signal.connect(self.focus_input)
-        self.controller.execute_on_main_thread_signal.connect(self.execute_callback)
         self.show()
 
+        # Start the assist function after a delay to prevent layout blip upon opening
         self.timer = QTimer.singleShot(500, self.delayed_setup)
 
     def delayed_setup(self):
@@ -295,19 +190,6 @@ class GUIWindow(QMainWindow):
             self.controller.current_input = user_input
             self.user_input_edit.clear()
             self.condition.notify()
-
-    def show_about_dialog(self):
-        version = os.environ.get("VERSION", "unknown")
-
-        message = f"""
-        Version {version}<br>
-        Copyright © 2024 Parsagon
-        """
-
-        # Title appears to be broken
-        box = QMessageBox()
-        box.about(self, "", message)
-        box.setWindowTitle("About Parsagon")
 
     @pyqtSlot(bool)
     def set_loading(self, loading):
@@ -325,6 +207,10 @@ class GUIWindow(QMainWindow):
     @pyqtSlot(int)
     def set_progress(self, progress):
         self.progress_bar.setValue(progress)
+
+    @pyqtSlot()
+    def focus_input(self):
+        self.user_input_edit.setFocus()
 
     @pyqtSlot(str, str, str)
     def update_text_ui(self, text, text_color="user", background_color="#33B1FF"):
@@ -417,12 +303,3 @@ class GUIWindow(QMainWindow):
         self.messages_layout.addSpacerItem(self.spacer)
 
         self.scroll_area.verticalScrollBar().setValue(self.scroll_area.verticalScrollBar().maximum())
-
-    @pyqtSlot()
-    def focus_input(self):
-        self.user_input_edit.setFocus()
-
-    @pyqtSlot(object, QEventLoop, ResultContainer)
-    def execute_callback(self, callback, loop, result_container):
-        result_container.value = callback()  # Store the result of the callback
-        loop.quit()  # Exit the event loop to unblock the background thread
